Use top level statements in samples
