--- samples/InMemoryBackgroundProcessing/VideoUploaded.cs
@@ -1,8 +1,0 @@
-﻿namespace InMemoryBackgroundProcessing;
-
-public class VideoUploaded
-{
-    public string? VideoId { get; set; }
-    public string? Url { get; set; }
-    public long SizeBytes { get; set; }
-}
