+++ samples/SimplePublisher/Program.cs
@@ -1,29 +1,23 @@
-﻿namespace SimplePublisher;
+﻿using SimplePublisher;
 
-public class Program
-{
-    public static void Main(string[] args)
-    {
-        CreateHostBuilder(args).Build().Run();
-    }
+var host = Host.CreateDefaultBuilder(args)
+               .ConfigureServices((hostContext, services) =>
+               {
+                   services.AddEventBus(builder =>
+                   {
+                       // Transport agnostic configuration
+                       builder.Configure(o =>
+                       {
+                           o.Naming.Scope = "dev"; // queues will be prefixed by 'dev'
+                           o.Naming.UseFullTypeNames = false;
+                       });
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-        Host.CreateDefaultBuilder(args)
-            .ConfigureServices((hostContext, services) =>
-            {
-                services.AddEventBus(builder =>
-                {
-                    // Transport agnostic configuration
-                    builder.Configure(o =>
-                    {
-                        o.Naming.Scope = "dev"; // queues will be prefixed by 'dev'
-                        o.Naming.UseFullTypeNames = false;
-                    });
+                       // Transport specific configuration
+                       builder.AddAzureQueueStorageTransport("UseDevelopmentStorage=true;");
+                   });
 
-                    // Transport specific configuration
-                    builder.AddAzureQueueStorageTransport("UseDevelopmentStorage=true;");
-                });
+                   services.AddHostedService<PublisherService>();
+               })
+               .Build();
 
-                services.AddHostedService<PublisherService>();
-            });
-}
+await host.RunAsync();
